% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_slope_criteria.R
\name{adjust_slope_criteria}
\alias{adjust_slope_criteria}
\title{Internal function which increases or decreases the slope criteria acting as a
threshold every two seconds since the end of the previous qrs complex detected
when either to many or none slope changes are detected.}
\usage{
adjust_slope_criteria(
  samples_since_qrs_start = NULL,
  samples_per_2s = NULL,
  num_slope = NULL,
  slope_crit,
  slope_crit_min,
  slope_crit_max,
  mid_detection_check = FALSE,
  max_slope_detected = NULL
)
}
\arguments{
\item{samples_since_qrs_start}{samples since the previous qrs complex was detected.}

\item{samples_per_2s}{constant value of the number of samples contained in two seconds.}

\item{num_slope}{number of slope changes detected since the previous qrs complex was identified.}

\item{slope_crit}{threshold used agaist which to compare the filter of the signal.}

\item{slope_crit_min}{constant value of the minimum possible value for the slope_crit}

\item{slope_crit_max}{constant value of the maximum possible value for the slope_crit}
}
\value{
New value for the slope criteria
}
\description{
Internal function which increases or decreases the slope criteria acting as a
threshold every two seconds since the end of the previous qrs complex detected
when either to many or none slope changes are detected.
}
