% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_variables_for_detection.R
\name{initialize_variables_for_detection}
\alias{initialize_variables_for_detection}
\title{Internal function which initializes the variables uses throught the qrs detection program.}
\usage{
initialize_variables_for_detection(threshold, from_sample, slope_crit_max)
}
\arguments{
\item{threshold}{minimum value to be achieved by the filtered signal to detect a slope change}

\item{from_sample}{initial sample to be analysed.}

\item{slope_crit_max}{maximum threshold value}
}
\value{
list with the different values used: slope_crit, current_sample_number, t_values, filter, num_slope, sign, max_slope_detected, samples_since_qrs_start, first_sample_of_qrs_complex, detection_window_countdown, annotations.
}
\description{
Internal function which initializes the variables uses throught the qrs detection program.
}
