% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_csv.R
\name{output_csv}
\alias{output_csv}
\title{Generates a .csv file with the annotations.}
\usage{
output_csv(annotations, file = "./output.csv")
}
\arguments{
\item{annotations}{dataframe containing the detected heartbeats and their annotations}

\item{file}{full file path and name.}
}
\description{
Generates a .csv file with the annotations.
}
