% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrs_detection.R
\name{qrs_detection}
\alias{qrs_detection}
\title{QRS detection on a single-channel ECG signal. This algorithm uses slope changes to
detect QRS complexes, and is based on the SQRS tool from the WFDB toolkit.}
\usage{
qrs_detection(
  signal_data,
  sampling_rate = 360,
  threshold = 200,
  from_sample = 0,
  to_sample = Inf,
  verbose = FALSE
)
}
\arguments{
\item{signal_data}{dataframe or list with the single-channel signal to be analyzed.}

\item{sampling_rate}{samples per second recorded.}

\item{threshold}{threshold to be applied to the filterd signal in order to consider it beign a slope change}

\item{from_sample}{start of the qrs detection in samples.}

\item{to_sample}{end of the qrs detection in samples.}
}
\value{
Dataframe with the annotated heartbeats. The reference is in sample numbers
}
\description{
Original reference:
Moody, G., Pollard, T., & Moody, B. (2022). WFDB Software Package (version 10.7.0).
PhysioNet. https://doi.org/10.13026/gjvw-1m31`
}
