% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_wfdb_212_sig_2.R
\name{read_bin_wfdb_212_sig_2}
\alias{read_bin_wfdb_212_sig_2}
\title{Internal function which reads a binary .dat file with format 212, containing
2 alternating signal samples.}
\usage{
read_bin_wfdb_212_sig_2(file, num_samples)
}
\arguments{
\item{file}{Full name of the .dat file.}

\item{num_samples}{Number of samples contained in the file for a single signal.}
}
\value{
A dataframe with the two signals, signal_1 and signal_2
}
\description{
Internal function which reads a binary .dat file with format 212, containing
2 alternating signal samples.
}
