% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header_wfdb.R
\name{read_header_wfdb}
\alias{read_header_wfdb}
\title{Parses a wfdb header file, which usually uses the extension .hea. It assumes
a certain structure for these files}
\usage{
read_header_wfdb(hea_file, verbose = FALSE)
}
\arguments{
\item{hea_file}{Full name of the file containing header information.}
}
\value{
A list containing two other lists with different information on the signals of the record, either general or specific per sample.
}
\description{
Parses a wfdb header file, which usually uses the extension .hea. It assumes
a certain structure for these files
}
