% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_signal.R
\name{read_signal}
\alias{read_signal}
\title{Reads a file containing ECG signal data}
\usage{
read_signal(
  file,
  signal_col = 1,
  has_head = FALSE,
  header_dat = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{full file name, including extension.}

\item{signal_col}{sets the index of the column where the signal data starts (starting on 1)}

\item{has_head}{value for csv files, defines whether the file has a header row}

\item{header_dat}{list containig two lists within with information on the header for dat files}
}
\value{
signal data for 2 signals, returned as a dataframe
}
\description{
Reads a file containing ECG signal data
}
