% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_detection.R
\name{slope_detection}
\alias{slope_detection}
\title{Internal function containing the logic for the slope changes detection and the
annotation of the possible heartbeats.}
\usage{
slope_detection(
  constants,
  num_slope,
  filter,
  slope_crit,
  detection_window_countdown,
  sign,
  samples_since_qrs_start,
  first_sample_of_qrs_complex,
  max_slope_detected,
  annotations,
  current_sample_number
)
}
\arguments{
\item{constants}{list containing the constants involved in the slope detection}

\item{num_slope}{number of slope changes detected}

\item{filter}{filtered signal values which to compare against the slope_crit or threshold}

\item{slope_crit}{threshold applied to the filtered signal.}

\item{detection_window_countdown}{time countdown in samples of the window available for the next qrs complex detection}

\item{sign}{sign of the current slope direction}

\item{samples_since_qrs_start}{samples since the previous qrs complex was detected.}

\item{first_sample_of_qrs_complex}{number of signals between the previous and current qrs complexes.}

\item{max_slope_detected}{maximum filtered value detected}

\item{annotations}{dataframe containing the annotations with the possible heartbeats and their clasiffication}

\item{current_sample_number}{number of the sample number since the start of the qrs detection, taking into account the original start sample number}
}
\value{
list with the variables modified in this function: num_slope, slope_crit, detection_window_countdown, sign, first_sample_of_qrs_complex, max_slope_detecte and annotations.
}
\description{
Internal function containing the logic for the slope changes detection and the
annotation of the possible heartbeats.
}
